+++ StreamChatSwiftUITests/Tests/ChatChannel/Suggestions/TestCommandsConfig.swift
@@ -23,13 +23,9 @@ public class TestCommandsConfig: CommandsConfig {
     public func makeCommandsHandler(
         with channelController: ChatChannelController
     ) -> CommandsHandler {
-        let userSearchController = ChatUserSearchController_Mock.mock(client: chatClient)
-        userSearchController.users_mock = Self.mockUsers
         let mentionsCommand = MentionsCommandHandler(
             channelController: channelController,
-            userSearchController: userSearchController,
-            commandSymbol: mentionsSymbol,
-            mentionAllAppUsers: false
+            commandSymbol: mentionsSymbol
         )
         let giphyCommand = GiphyCommandHandler(commandSymbol: "/giphy")
         let muteCommand = MuteCommandHandler(